+++ Modules/32-Deduplication.ps1
@@ -40,6 +40,7 @@ function Show-DeduplicationManagement {
     }
 
     while ($true) {
+        if ($global:ReturnToMainMenu) { return }
         # Show current dedup status
         Clear-Host
         Write-OutputColor "" -color "Info"